+++ packages/featureserver/src/geometry/index.js
@@ -1,0 +1,5 @@
+const computeSpatialReference = require('./spatialReference')
+const computeExtent = require('./extents')
+const geometryMap = require('./geometry-map')
+
+module.exports = { computeSpatialReference, computeExtent, geometryMap }