+++ packages/featureserver/src/templates.js
@@ -1,15 +1,16 @@
 const _ = require('lodash')
 const moment = require('moment')
-const Utils = require('./utils')
-const field = require('./field')
+const { isTable } = require('./utils')
+const { computeFieldObject, createFieldAliases, createStatFields } = require('./field')
+const { computeSpatialReference, computeExtent } = require('./geometry')
 
-module.exports = { render, renderServer, renderStatistics }
+module.exports = { renderLayer, renderFeatures, renderStatistics, renderServer, renderStats }
 
 const templates = {
-  server: require('../templates/server.json'),
   layer: require('../templates/layer.json'),
   features: require('../templates/features.json'),
   statistics: require('../templates/statistics.json'),
+  server: require('../templates/server.json'),
   field: require('../templates/field.json'),
   objectIDField: require('../templates/oid-field.json')
 }
@@ -20,18 +21,15 @@ const renderers = {
   esriGeometryPoint: require('../templates/renderers/point.json')
 }
 
-const defaultSR = { wkid: 4326 }
-const mercatorSR = { wkid: 102100, latestWkid: 3857 }
-
 /**
- * loads a template json file and attaches fields
+ * loads a template layer json file and attaches fields
  *
- * @param {string} tmpl
- * @param {object} data
+ * @param {object} featureCollection
+ * @param {object} options
  * @return {object} template
  */
-function render (template, featureCollection = {}, options = {}) {
-  const json = _.cloneDeep(templates[template])
+function renderLayer (featureCollection = {}, options = {}) {
+  const json = _.cloneDeep(templates.layer)
   const data = featureCollection
   const metadata = data.metadata || {}
   if (!json) throw new Error('Unsupported operation')
@@ -40,19 +38,41 @@ function render (template, featureCollection = {}, options = {}) {
   if (json.fullExtent) json.fullExtent = json.initialExtent = json.extent = metadata.extent || options.extent
   else if (json.extent) json.extent = metadata.extent || options.extent
 
+  json.id = parseInt(options.layer) || 0
   if (json.geometryType) json.geometryType = options.geometryType
   if (json.spatialReference) json.spatialReference = computeSpatialReference(options.spatialReference)
   if (json.name && metadata.name) json.name = metadata.name
   if (json.description && metadata.description) json.description = metadata.description
   if (json.extent && metadata.extent) json.extent = computeExtent(metadata.extent)
-  if (json.features) json.features = data.features
-  if (json.fields) json.fields = field.computeFieldObject(data, template, options)
-  if (json.type) json.type = Utils.isTable(json, data) ? 'Table' : 'Feature Layer'
+  if (json.fields) json.fields = computeFieldObject(data, 'layer', options)
+  if (json.type) json.type = isTable(json, data) ? 'Table' : 'Feature Layer'
   if (json.drawingInfo) json.drawingInfo.renderer = renderers[json.geometryType]
   if (json.timeInfo) json.timeInfo = metadata.timeInfo
   if (json.maxRecordCount) json.maxRecordCount = metadata.maxRecordCount || 1000
   if (json.displayField) json.displayField = metadata.displayField || json.fields[0].name
   if (json.objectIdField) json.objectIdField = metadata.idField || 'OBJECTID'
+  return json
+}
+
+function renderFeatures (featureCollection = {}, options = {}) {
+  const json = _.cloneDeep(templates.features)
+  const data = featureCollection
+  if (!json) throw new Error('Unsupported operation')
+
+  if (json.geometryType) json.geometryType = options.geometryType
+  if (json.spatialReference) json.spatialReference = computeSpatialReference(options.spatialReference)
+  if (json.fields) json.fields = computeFieldObject(data, 'layer', options)
+  if (json.features) json.features = data.features
+  return json
+}
+
+function renderStatistics (featureCollection = {}, options = {}) {
+  const json = _.cloneDeep(templates.statistics)
+  const data = featureCollection
+  if (!json) throw new Error('Unsupported operation')
+
+  if (json.fields) json.fields = computeFieldObject(data, 'statistics', options)
+  if (json.features) json.features = data.features
   return json
 }
 
@@ -67,80 +87,16 @@ function renderServer (server, { layers, tables }) {
   return json
 }
 
-function renderStatistics (data) {
+function renderStats (data) {
   let stats = data.statistics
   if (!Array.isArray(stats)) stats = [stats]
-  const fields = data.metadata ? field.computeFieldObject(data) : createStatFields(stats)
+  const fields = data.metadata ? computeFieldObject(data) : createStatFields(stats)
   return {
     displayFieldName: '',
     fieldAliases: createFieldAliases(stats),
     fields,
     features: createStatFeatures(stats)
   }
-}
-
-function computeSpatialReference (sr) {
-  if (!sr) return defaultSR
-  else if (sr === 4326 || sr.wkid === 4326 || sr.latestWkid === 4326) return defaultSR
-  else if (sr === 102100 || sr === 3857 || sr.wkid === 102100 || sr.latestWkid === 3857) return mercatorSR
-  else if (typeof sr === 'number') return { wkid: sr }
-  else {
-    return {
-      wkid: sr.wkid || sr.latestWkid,
-      latestWkid: sr.latestWkid || sr.wkid
-    }
-  }
-}
-
-function computeExtent (input) {
-  let coords
-  if (input.xmin) return input
-  if (Array.isArray(input)) {
-    if (Array.isArray(input[0])) coords = input
-    else coords = [[input[0], input[1]], [input[2], input[3]]]
-  } else {
-    throw new Error('invalid extent passed in metadata')
-  }
-  return {
-    xmin: coords[0][0],
-    ymin: coords[0][1],
-    xmax: coords[1][0],
-    ymax: coords[1][1],
-    spatialReference: {
-      wkid: 4326,
-      latestWkid: 4326
-    }
-  }
-}
-
-function createFieldAliases (stats) {
-  const fields = Object.keys(stats[0])
-  return fields.reduce((aliases, field) => {
-    aliases[field] = field
-    return aliases
-  }, {})
-}
-
-function createStatFields (stats) {
-  return Object.keys(stats[0]).map(field => {
-    const sample = _.find(stats, s => {
-      return stats[field] !== null
-    })
-    const statField = {
-      name: field,
-      type: detectType(sample[field]),
-      alias: field
-    }
-    if (statField.type === 'esriFieldTypeString') statField.length = 254
-    return statField
-  }, {})
-}
-
-function detectType (value) {
-  if (!value) return null
-  else if (moment(value, [moment.ISO_8601], true).isValid()) return 'esriFieldTypeDate'
-  else if (typeof value === 'string') return 'esriFieldTypeString'
-  else if (typeof value === 'number') return 'esriFieldTypeDouble'
 }
 
 function createStatFeatures (stats) {