+++ packages/featureserver/CHANGELOG.md
@@ -2,6 +2,14 @@
 All notable changes to this project will be documented in this file.
 This project adheres to [Semantic Versioning](http://semver.org/).
 
+
+## Unreleased
+### Fixed
+* Save id property when rendering layer
+### Changed
+* Reorganize templates.js into individual render functions
+* Renamed renderStatistics to renderStats for render function consistency
+
 ## [2.3.7] - 2017-07-06
 ### Fixed
 - supportsPagination => true